move tests to be with other marker gene tests
