+++ frontend/tests/features/wheresMyGene/cellTooltip.test.ts
@@ -7,15 +7,7 @@ import {
   CELL_TYPE_ROW_CLASS_NAME,
 } from "src/views/WheresMyGeneV2/components/HeatMap/components/YAxisChart/constants";
 import { test } from "tests/common/test";
-import assert from "assert";
-import {
-  NO_MARKER_GENES_DESCRIPTION,
-  NO_MARKER_GENES_FOR_BLOOD_DESCRIPTION,
-  TOO_FEW_CELLS_NO_MARKER_GENES_DESCRIPTION,
-} from "src/views/WheresMyGeneV2/components/CellInfoSideBar/constants";
 const { describe } = test;
-
-const NO_MARKER_GENES_DESCRIPTION_ID = "no-marker-genes-description";
 
 describe("cell tooltip", () => {
   test(`Should verify cell tooltip hover`, async ({ page }) => {
@@ -52,67 +44,5 @@ describe("cell tooltip", () => {
 
       checkedCells++;
     }
-  });
-  test(`Should verify blood cells have no marker genes`, async ({ page }) => {
-    await goToPage(`${TEST_URL}${ROUTES.WHERE_IS_MY_GENE}`, page);
-
-    // Expand blood tissue
-    await expandTissue(page, "blood");
-
-    // Click stem cell info icon
-    await page.getByTestId("cell-type-info-button-blood-stem cell").click();
-
-    // Verify copy is what we expect
-    const noMarkerGenesDescription = (await page
-      .getByTestId(NO_MARKER_GENES_DESCRIPTION_ID)
-      .textContent()) as string;
-    assert.strictEqual(
-      noMarkerGenesDescription.trim(),
-      NO_MARKER_GENES_FOR_BLOOD_DESCRIPTION
-    );
-  });
-  test(`Should verify cell types with < 25 cells have no marker genes`, async ({
-    page,
-  }) => {
-    await goToPage(`${TEST_URL}${ROUTES.WHERE_IS_MY_GENE}`, page);
-
-    // Expand blood tissue
-    await expandTissue(page, "adipose-tissue");
-
-    // Click naive B cell info icon
-    await page
-      .getByTestId("cell-type-info-button-adipose tissue-naive B cell")
-      .click();
-
-    // Verify copy is what we expect
-    const noMarkerGenesDescription = (await page
-      .getByTestId(NO_MARKER_GENES_DESCRIPTION_ID)
-      .textContent()) as string;
-    assert.strictEqual(
-      noMarkerGenesDescription.trim(),
-      TOO_FEW_CELLS_NO_MARKER_GENES_DESCRIPTION
-    );
-  });
-  test(`Should verify copy for cell types with no marker genes`, async ({
-    page,
-  }) => {
-    await goToPage(`${TEST_URL}${ROUTES.WHERE_IS_MY_GENE}`, page);
-
-    // Expand blood tissue
-    await expandTissue(page, "yolk-sac");
-
-    // Click yolk sac somatic cell info icon
-    await page
-      .getByTestId("cell-type-info-button-yolk sac-somatic cell")
-      .click();
-
-    // Verify copy is what we expect
-    const noMarkerGenesDescription = (await page
-      .getByTestId(NO_MARKER_GENES_DESCRIPTION_ID)
-      .textContent()) as string;
-    assert.strictEqual(
-      noMarkerGenesDescription.trim(),
-      NO_MARKER_GENES_DESCRIPTION
-    );
   });
 });
